Changed default variables to float instead of double because JOGL uses floats... for compability reasons!
//Dan

+++ src/pong/model/Ball.java
@@ -2,7 +2,7 @@ package pong.model;
 
 public class Ball extends MoveableItem {
 
-	public Ball(double xPos, double yPos, double zPos) {
+	public Ball(float xPos, float yPos, float zPos) {
 		super(xPos, yPos, zPos);
 	}
