+++ src/pong/model/MoveableItem.java
@@ -2,7 +2,7 @@ package pong.model;
 
 public class MoveableItem extends GameItem{
 
-	public MoveableItem(double xPos, double yPos, double zPos) {
+	public MoveableItem(float xPos, float yPos, float zPos) {
 		super(xPos, yPos, zPos);
 	}
